Fix binary encoding issues
- The binary reader/writer classes currently only read/write 0-8 bits at a time. Encoding large event 'n' values easily exceeds the 8 bit limit, so it must account for this limitation by breaking 'n' into chucks. Eventually, the readers/writers should support 32 bits natively, but adjusting the encoder will work for now.
- Test cases of 513 (B=9), 21474836 (B=24), and int.MaxValue all exposed issues, so the unit tests remain.

+++ src/Itc4net/Binary/BitProcessor.cs
@@ -21,7 +21,7 @@ namespace Itc4net.Binary
         }
 
         protected Stream BaseStream => _stream;
-        protected byte RelativePosition { get; set; }
+        protected int RelativePosition { get; set; }
         protected int CurrentByte { get; set; }
 
         /// <summary>
